+++ .idea/workspace.xml
@@ -1,0 +1,95 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ChangeListManager">
+    <list default="true" id="05db42f5-68c7-4b38-a7f9-19166bf4f71a" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Shared/Hooks/useChangeDependency.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Shared/Hooks/useComponentRect.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Shared/Hooks/useForceRendering.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Shared/Hooks/useIsShowing.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Shared/Hooks/useSpecificTimeEffect.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Shared/Hyundux/Store.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/Caecae/src/Shared/Hyundux/Store.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Shared/Types/FindingGameAnswer.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Shared/Util/FindChildrenElement.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Shared/Util/areEqual.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Widget/Footer/Footer.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Widget/InfoSection/InfoSection.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Widget/LottieContainer/LottieContainter.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Widget/Navigation/Navigation.css" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Widget/Navigation/Navigation.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Widget/Overlay/Overlay.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Widget/Overlay/OverlayContent.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Widget/PictureGameBoard/PictureGameBoard.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Widget/Shape/Circle.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/Widget/SmileBadge/SmileBadge.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/jobs/FindingGame/FindingGame.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/Caecae/src/jobs/FindingGame/FindingGame.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/EventInfoLanding/Sections/GoToEvent.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/EventInfoLanding/Sections/ShowCasper1.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/EventInfoLanding/Sections/ShowCasper2.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/EventInfoLanding/Sections/ShowCasper3.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/EventInfoLanding/Sections/ShowCasper4.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/FindingGameLanding/FindingGameLadingPage.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/Caecae/src/pages/FindingGameLanding/FindingGameLadingPage.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/FindingGameLanding/Sections/EventPeriod.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/FindingGameLanding/Sections/HowToEvent.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/FindingGameLanding/Sections/LadingPageTitle.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/FindingGameLanding/Sections/OpenEvent.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/FindingGameLanding/Sections/SelectionMethod.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/RacingGameLanding/RacingGameLandingPage.tsx" beforeDir="false" afterPath="$PROJECT_DIR$/Caecae/src/pages/RacingGameLanding/RacingGameLandingPage.tsx" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/RacingGameLanding/Sections/EventIntro.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/RacingGameLanding/Sections/EventPeriod.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/RacingGameLanding/Sections/EventPrecaution.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/RacingGameLanding/Sections/GiftInfo.tsx" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/Caecae/src/pages/RacingGameLanding/Sections/HowToEvent.tsx" beforeDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProjectColorInfo"><![CDATA[{
+  "customColor": "",
+  "associatedIndex": 7
+}]]></component>
+  <component name="ProjectId" id="2kOyxwPGxVQ66ucm14GR54wfQO3" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "RunOnceActivity.OpenProjectViewOnStart": "true",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "git-widget-placeholder": "CC-131",
+    "kotlin-language-version-configured": "true",
+    "last_opened_file_path": "/Users/changgyoseo/Projects/caecae",
+    "nodejs_package_manager_path": "yarn",
+    "vue.rearranger.settings.migration": "true"
+  }
+}]]></component>
+  <component name="ReactDesignerToolWindowState">
+    <option name="myId2Visible">
+      <map>
+        <entry key="com.intellij.reactbuddy.reactComponents" value="false" />
+        <entry key="com.intellij.reactbuddy.reactInspector" value="false" />
+        <entry key="com.intellij.reactbuddy.storybook" value="false" />
+      </map>
+    </option>
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="05db42f5-68c7-4b38-a7f9-19166bf4f71a" name="Changes" comment="" />
+      <created>1723169169643</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1723169169643</updated>
+      <workItem from="1723169170774" duration="9000" />
+    </task>
+    <servers />
+  </component>
+</project>
