+++ Caecae/src/components/common/SmileBadge/SmileBadge.tsx
