+++ Caecae/src/hooks/useIsShowing.tsx
