+++ Caecae/src/jobs/FindingGame/FindingGame.tsx
@@ -2,7 +2,7 @@ import { createState } from "../../Shared/Hyundux/State";
 import { makePayLoad } from "../../Shared/Hyundux/Util/StoreUtil";
 import Reducer from "../../Shared/Hyundux/Reducer";
 import { Action } from "../../Shared/Hyundux/Actions";
-import FindingGameAnswer from "../../Shared/Types/FindingGameAnswer";
+import FindingGameAnswer from "../../types/FindingGameAnswer";
 
 const WORK_NAME = "FindingGame";
