+++ Caecae/src/pages/RacingGameLanding/RacingGameLandingPage.tsx
@@ -1,9 +1,9 @@
 import Navigation from "../../Widget/Navigation/Navigation";
-import EventIntro from "./Sections/EventIntro";
-import HowToEvent from "./Sections/HowToEvent";
-import GiftInfo from "./Sections/GiftInfo";
-import EventPeriod from "./Sections/EventPeriod";
-import EventPrecaution from "./Sections/EventPrecaution";
+import EventIntro from "./RacingGameLading/EventIntro";
+import HowToEvent from "./RacingGameLading/HowToEvent";
+import GiftInfo from "./RacingGameLading/GiftInfo";
+import EventPeriod from "./RacingGameLading/EventPeriod";
+import EventPrecaution from "./RacingGameLading/EventPrecaution";
 import Footer from "../../Widget/Footer/Footer";
 
 const RacingGameLandingPage = () => {
